+++ CHANGELOG.md
@@ -5,27 +5,41 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+
 ## [Unreleased]
 
 ### Added
 
+### Changed
 
-- Processes that take a CRS as argument now try harder to convert your input into a proper EPSG code, to avoid unexpected results when an invalid argument gets sent to the backend.
+### Removed
+
+### Fixed
+
+
+
+## [0.22.0] - 2023-08-09
+
+### Added
+
+- Processes that take a CRS as argument now try harder to normalize your input to
+  a CRS representation that aligns with the openEO API (using `pyproj` library when available)
+  ([#259](https://github.com/Open-EO/openeo-python-client/issues/259))
 - Initial `load_geojson` support with `Connection.load_geojson()` ([#424](https://github.com/Open-EO/openeo-python-client/issues/424))
 - Initial `load_url` (for vector cubes) support with `Connection.load_url()` ([#424](https://github.com/Open-EO/openeo-python-client/issues/424))
+- Add `VectorCube.apply_dimension()` ([Open-EO/openeo-python-driver#197](https://github.com/Open-EO/openeo-python-driver/issues/197))
 - Support lambda based property filtering in `Connection.load_stac()` ([#425](https://github.com/Open-EO/openeo-python-client/issues/425))
-
+- `VectorCube`: initial support for `filter_bands`, `filter_bbox`, `filter_labels` and `filter_vector` ([#459](https://github.com/Open-EO/openeo-python-client/issues/459))
 
 ### Changed
 
 - `Connection` based requests: always use finite timeouts by default (20 minutes in general, 30 minutes for synchronous execute requests)
   ([#454](https://github.com/Open-EO/openeo-python-client/issues/454))
 
-### Removed
-
 ### Fixed
 
 - Fix: MultibackendJobManager should stop when finished, also when job finishes with error ([#452](https://github.com/Open-EO/openeo-python-client/issues/432))
+
 
 ## [0.21.1] - 2023-07-19